Prevent authentication callback from being reused
Within the MSA flow there are scenarios where an authentication callback
would recieve a failure, and then a success message due to how slient
authentication is treated during an interactive login.  Changing the SDK
to handle this scenario and isolate this strange case so callbacks should
only be acted on once.