Use AcquireToken to force refresh token to be saved in the cache
Fixes #36
When using the AquireTokenByRefreshToken(...) method, the underlying ADAL
system does not guarentee the token will be saved for future use in the
TokenCacheStore.  Tracing into the AuthenticationContext implementation,
refreshToken(..., final boolean useCache) is getting set to false in all
of the 'refresh token` scenarios.  Using aquireToken and then depending on
the Prompt settings seems to be the best course of action to mitigate this
issue.  Note: it is not clear why this behavoir changed.