Handle error responses that are not json in nature
Some error messages that can comeback from the service can be in non-json
formats.  Malformed URLs are an example of an error that can return this
kind of message.  Instead of assuming all messages are json, instead check
the content-type header and write the raw error message info on toString
for the exception details.

Fixes #53